% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsmngr-package.R
\docType{package}
\name{contactsmngr-package}
\alias{contactsmngr}
\alias{contactsmngr-package}
\title{contactsmngr: Contacts manager}
\description{
Organize your contacts using an RStudio addin
}
\author{
\strong{Maintainer}: Philipp Baessler \email{mail@philippb.com}

}
\keyword{internal}
