% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contact.R
\name{create_contact}
\alias{create_contact}
\title{Title}
\usage{
create_contact(
  id = paste(surname, forename, sep = "_"),
  surname,
  forename,
  middle_name = NULL,
  used_name = forename,
  birthday = NULL,
  siezen = NULL,
  business_emails,
  private_emails = NULL,
  business_phone = NULL,
  private_phone = NULL,
  `function` = NULL,
  employment = NULL,
  first_encounter = NULL,
  expertise = NULL,
  points = 0,
  mode = c("check", "add", "overwrite")
)
}
\arguments{
\item{mode}{}
}
\description{
Title
}
