% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contactsDetails.R
\name{contactsDetailsUI}
\alias{contactsDetailsUI}
\alias{contactsDetailsServer}
\title{Title}
\usage{
contactsDetailsUI(id)

contactsDetailsServer(id)
}
\arguments{
\item{id}{}
}
\description{
Title
}
\keyword{internal}
