% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_contactsList.R
\name{contactsListUI}
\alias{contactsListUI}
\alias{contactsListServer}
\title{Title}
\usage{
contactsListUI(id)

contactsListServer(id)
}
\arguments{
\item{id}{}
}
\description{
Title
}
\keyword{internal}
